+++ fhir/fhir-release-2/src/resources/allergy_intolerance.rs
@@ -64,7 +64,7 @@ pub struct AllergyIntolerance {
 
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r2::resources::Resource>,
 
     /// Additional Content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]