+++ fhir/fhir-release-3/src/resources/audit_event.rs
@@ -57,7 +57,7 @@ pub struct AuditEvent {
 
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r3::resources::Resource>,
 
     /// Additional Content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]