+++ fhir/fhir-release-3/src/resources/endpoint.rs
@@ -58,7 +58,7 @@ pub struct Endpoint {
 
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r3::resources::Resource>,
 
     /// Additional Content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]