+++ fhir/fhir-release-4/src/resources/document_reference.rs
@@ -61,7 +61,7 @@ pub struct DocumentReference {
 
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r4::resources::Resource>,
 
     /// Additional content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]