+++ fhir/fhir-release-4/src/resources/medicinal_product_packaged.rs
@@ -57,7 +57,7 @@ pub struct MedicinalProductPackaged {
 
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r4::resources::Resource>,
 
     /// Additional content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]