+++ fhir/fhir-release-4/src/resources/practitioner.rs
@@ -66,7 +66,7 @@ pub struct Practitioner {
 
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r4::resources::Resource>,
 
     /// Additional content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]