+++ fhir/fhir-release-5/src/resources/explanation_of_benefit.rs
@@ -89,7 +89,7 @@ pub struct ExplanationOfBenefit {
     pub text: Option<types::Narrative>,
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r5::resources::Resource>,
     /// Additional content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub extension: Vec<types::Extension>,