+++ fhir/fhir-release-5/src/resources/insurance_plan.rs
@@ -86,7 +86,7 @@ pub struct InsurancePlan {
 
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r5::resources::Resource>,
 
     /// Additional content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]