+++ fhir/fhir-release-5/src/resources/terminology_capabilities.rs
@@ -91,7 +91,7 @@ pub struct TerminologyCapabilities {
 
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r5::resources::Resource>,
 
     /// Additional content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]