+++ fhir/fhir-release-6/src/resources/diagnostic_report.rs
@@ -71,7 +71,7 @@ pub struct DiagnosticReport {
 
     /// Contained, inline Resources
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub contained: Vec<::serde_json::Value>,
+    pub contained: Vec<crate::r6::resources::Resource>,
 
     /// Additional content defined by implementations
     #[serde(default, skip_serializing_if = "Vec::is_empty")]